# Conflicts:
#	dtu-pay-E2ETest/src/test/java/dtu/pay/CommonUserSteps.java

+++ payment-service/src/main/java/payment/service/models/PaymentRecord.java
@@ -1,5 +1,5 @@
 package payment.service.models;
 
-public record PaymentRecord(int amount, String token, String customerId, String merchantId)
-{
-}
+import java.math.BigDecimal;
+
+public record PaymentRecord(BigDecimal amount, String token, String customerId, String merchantId) {}
